Cleaning up

+++ src/main/java/com/hotel/assignment/dto/PaymentUpdateEvent.java
@@ -1,8 +1,10 @@
 package com.hotel.assignment.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentUpdateEvent(
         String paymentId,
         String debtorAccountnumber,
-        Double amountReceived,
+        BigDecimal amountReceived,
         String transactionDescription
 ) {}
